Reorganize template functions into template package
* Move all functions used in go templates to the template package.
* Move functions only used in one file from deploy/cloudformation/
  stack/template_functions.go to deploy/cloudformation/stack/svc.go
* Rename LogicalIDSafe used in mft/cf to ReplaceDashes (as the behavior
  is quite different from the LogicalIDSafe function used for storage
  environment variables)
	modified:   internal/pkg/deploy/cloudformation/stack/app.go
