+++ internal/pkg/template/service_test.go
@@ -74,32 +74,6 @@ func TestTemplate_ParseSvc(t *testing.T) {
 	}
 }
 
-func TestToSnakeCase(t *testing.T) {
-	testCases := map[string]struct {
-		in     string
-		wanted string
-	}{
-		"camel case: starts with uppercase": {
-			in:     "AdditionalResourcesPolicyArn",
-			wanted: "ADDITIONAL_RESOURCES_POLICY_ARN",
-		},
-		"camel case: starts with lowercase": {
-			in:     "additionalResourcesPolicyArn",
-			wanted: "ADDITIONAL_RESOURCES_POLICY_ARN",
-		},
-		"all lower case": {
-			in:     "myddbtable",
-			wanted: "MYDDBTABLE",
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			require.Equal(t, tc.wanted, ToSnakeCase(tc.in))
-		})
-	}
-}
-
 func TestHasSecrets(t *testing.T) {
 	testCases := map[string]struct {
 		in     ServiceOpts
